Revert "Map tooltip"

+++ plot-builder/src/commonMain/kotlin/jetbrains/datalore/plot/builder/layout/PlotLayoutBase.kt
@@ -6,13 +6,13 @@
 package jetbrains.datalore.plot.builder.layout
 
 internal abstract class PlotLayoutBase : PlotLayout {
-    protected var paddingTop: Double = 0.0
+    protected var paddingTop: Double = 0.toDouble()
         private set
-    protected var paddingRight: Double = 0.0
+    protected var paddingRight: Double = 0.toDouble()
         private set
-    protected var paddingBottom: Double = 0.0
+    protected var paddingBottom: Double = 0.toDouble()
         private set
-    protected var paddingLeft: Double = 0.0
+    protected var paddingLeft: Double = 0.toDouble()
         private set
 
     override fun setPadding(top: Double, right: Double, bottom: Double, left: Double) {
